+++ recipes/database.rb
@@ -19,11 +19,6 @@
 
 settings = merge_confluence_settings
 
-database_connection = {
-  :host => settings['database']['host'],
-  :port => settings['database']['port']
-}
-
 case settings['database']['type']
 when 'mysql'
   mysql2_chef_gem 'confluence' do
@@ -40,10 +35,8 @@ when 'mysql'
     action [:create, :start]
   end
 
-  database_connection.merge!(:username => 'root', :password => node['mysql']['server_root_password'])
-
   mysql_database settings['database']['name'] do
-    connection database_connection
+    connection confluence_database_connection
     collation 'utf8_bin'
     encoding 'utf8'
     action :create
@@ -51,13 +44,13 @@ when 'mysql'
 
   # See this MySQL bug: http://bugs.mysql.com/bug.php?id=31061
   mysql_database_user '' do
-    connection database_connection
+    connection confluence_database_connection
     host 'localhost'
     action :drop
   end
 
   mysql_database_user settings['database']['user'] do
-    connection database_connection
+    connection confluence_database_connection
     host '%'
     password settings['database']['password']
     database_name settings['database']['name']
@@ -67,16 +60,15 @@ when 'mysql'
 when 'postgresql'
   include_recipe 'postgresql::server'
   include_recipe 'database::postgresql'
-  database_connection.merge!(:username => 'postgres', :password => node['postgresql']['password']['postgres'])
 
   postgresql_database_user settings['database']['user'] do
-    connection database_connection
+    connection confluence_database_connection
     password settings['database']['password']
     action :create
   end
 
   postgresql_database settings['database']['name'] do
-    connection database_connection
+    connection confluence_database_connection
     connection_limit '-1'
     encoding 'utf8'
     owner settings['database']['user']